updated for version 7.3.1305 Problem:    Warnings from 64 bit compiler. Solution:   Add type casts.

+++ src/misc2.c
@@ -4680,8 +4680,9 @@ vim_findfile_init(path, filename, stopdirs, level, free_visited, find_what,
     STRCPY(ff_expand_buffer, search_ctx->ffsc_start_dir);
     add_pathsep(ff_expand_buffer);
     {
-	int    eb_len = STRLEN(ff_expand_buffer);
-	char_u *buf = alloc(eb_len + STRLEN(search_ctx->ffsc_fix_path) + 1);
+	int    eb_len = (int)STRLEN(ff_expand_buffer);
+	char_u *buf = alloc(eb_len
+				+ (int)STRLEN(search_ctx->ffsc_fix_path) + 1);
 
 	STRCPY(buf, ff_expand_buffer);
 	STRCPY(buf + eb_len, search_ctx->ffsc_fix_path);
@@ -4700,18 +4701,18 @@ vim_findfile_init(path, filename, stopdirs, level, free_visited, find_what,
 
 	    if (p > search_ctx->ffsc_fix_path)
 	    {
-		len = p - search_ctx->ffsc_fix_path - 1;
+		len = (int)(p - search_ctx->ffsc_fix_path) - 1;
 		STRNCAT(ff_expand_buffer, search_ctx->ffsc_fix_path, len);
 		add_pathsep(ff_expand_buffer);
 	    }
 	    else
-		len = STRLEN(search_ctx->ffsc_fix_path);
+		len = (int)STRLEN(search_ctx->ffsc_fix_path);
 
 	    if (search_ctx->ffsc_wc_path != NULL)
 	    {
 		wc_path = vim_strsave(search_ctx->ffsc_wc_path);
-		temp = alloc(STRLEN(search_ctx->ffsc_wc_path)
-				 + (STRLEN(search_ctx->ffsc_fix_path) - len));
+		temp = alloc((int)(STRLEN(search_ctx->ffsc_wc_path)
+				 + (STRLEN(search_ctx->ffsc_fix_path)) - len));
 	    }
 
 	    if (temp == NULL || wc_path == NULL)
